requirements.txt
biopython == 1.79
requests==2.9.1
python-dotenv